+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * d028084 - 2017-09-18: Release v9.1.18 
+- * f692c2a - 2017-09-27: Updated HTBits provider to use https for url 
+- * 4b1e431 - 2017-09-18: Release v9.1.18 
 - * a6a84d4 - 2017-09-18: Fixed issue with retrieving cached results for searches when performing a provider search 
 - * 1e5bd13 - 2017-09-11: Fix for issue #71 - removes format metadata from video file 
 - * 906b7df - 2017-09-10: Fixed issue #78 - ASCII error while manual postprocessing